Refactor and extend tests
- split up tests in multiple files
- add tests
- add tests which fail as soon as the corresponding functions are implemented
